requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
flask-restx==1.3.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-decouple==3.8
pytz==2024.1
referencing==0.33.0
rpds-py==0.17.1
Werkzeug==3.0.1
